requirements.txt
# Google AI and Gemini API
google-generativeai>=0.3.0

# API and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scrapy>=2.11.0
aiohttp>=3.8.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Text processing and NLP
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0
sentence-transformers>=2.2.0

# API frameworks
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
# Note: sqlite3 is part of Python standard library

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.28.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Additional packages for fine-tuning integration
tensorboard>=2.14.0  # For training monitoring
wandb>=0.16.0  # Alternative training monitoring (optional)
einops>=0.7.0  # For tensor operations
